cope with nonexisting root logs
There's a rare chance a repository gets misconfigured or whatever and a package that should not be available on a specific architecture is available.

For example, the 'seabios' SRPM in Fedora ELN only builds for x86_64. But it builds some 'noarch' packages, too. When we enabled the buildroot repo for ELN, it has these noarch available everywhere. This causes Content Resolver to look for their root logs in those architectures, but they don't exist.