Do not inherit repositories in addons
An addon repository does not necessarily want the same repos or exclusions
as its base, and vice versa.  For example, ELN should really not include
Extras, and a lot of unwanted packages that we want to filter from
Buildroot or Rawhide may still be eligible for ELN Extras.

v2: the new repo must be listed in an environment, otherwise no packages
are found.